# perf: add `TrimStart(this StringBuilder`