Prevents duplication of the result string, saving memory in certain cases, ie. a preprocessor directive at the start of a file.

#### Benchmarks (timing is inaccurate didn't bother rerunning it)
### Before
| Method                        | Mean      | Error    | StdDev   | Gen0      | Gen1      | Allocated |
|------------------------------ |----------:|---------:|---------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   |  98.32 ms | 1.954 ms | 2.327 ms | 2000.0000 | 1000.0000 |  31.07 MB |
| Default_CodeFormatter_Complex | 206.40 ms | 3.671 ms | 5.716 ms | 5000.0000 | 2000.0000 |  54.58 MB |

### After
| Method                        | Mean      | Error    | StdDev   | Gen0      | Gen1      | Allocated |
|------------------------------ |----------:|---------:|---------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   |  86.67 ms | 1.675 ms | 1.929 ms | 2000.0000 | 1000.0000 |  31.07 MB |
| Default_CodeFormatter_Complex | 182.89 ms | 3.469 ms | 6.766 ms | 5000.0000 | 2000.0000 |  52.47 MB |