@TimothyMakkison 2025-03-21
> FYI I'm heading to the airport real soon so won't be around for the next 8 days. I may be able to get caught up before I take off.

Happy journeys, hope you have a good holiday 😄 


> Hoping to release all of these performance changes along with 1.0 in the next couple weeks, thanks again for all your hard work!

Looking forward to it. I'm near the end of my changes. See #1551 for a general feel for the final result. I'll have to find a way to benchmark an actual run of multiple files, to see if I've overlooked something.